♻️ Refactor run approval notification, restore old server.js approval run
