+++ src/routes/ngs-api/approveRun/+server.js
@@ -2,6 +2,7 @@ import sql from 'mssql';
 import { json } from '@sveltejs/kit';
 
 import * as dotenv from 'dotenv';
+import { notifyDiscordNewRunApproved } from '$lib/server/discordNotify';
 
 dotenv.config();
 
@@ -130,7 +131,7 @@ export async function POST({ request }) {
 				poolConnection.close();
 
 				//returner = context.req.body;
-				notifyDiscord('<Moderator_Name>', '<Player_Name>', 'DFA Duo');
+				notifyDiscordNewRunApproved('<Moderator_Name>', '<Player_Name>', 'DFA Duo');
 				return json({ Code: 'success' });
 			} catch (err) {
 				console.error(err.message);
@@ -304,7 +305,7 @@ export async function POST({ request }) {
 				poolConnection.close();
 
 				//returner = context.req.body;
-				notifyDiscord('<Moderator_Name>', '<Player_Name>', 'DFA Party');
+				notifyDiscordNewRunApproved('<Moderator_Name>', '<Player_Name>', 'DFA Party');
 				return json({ Code: 'success' });
 			} catch (err) {
 				console.error(err.message);
@@ -434,7 +435,7 @@ export async function POST({ request }) {
 
 				//returner = context.req.body;
 
-				notifyDiscord('<Moderator_Name>', '<Player_Name>', 'Purple Party');
+				notifyDiscordNewRunApproved('<Moderator_Name>', '<Player_Name>', 'Purple Party');
 				return json({ Code: 'success' });
 			} catch (err) {
 				console.error(err.message);
@@ -544,7 +545,7 @@ export async function POST({ request }) {
 				poolConnection.close();
 
 				//returner = context.req.body;
-				notifyDiscord('<Moderator_Name>', '<Player_Name>', 'Purple Duo');
+				notifyDiscordNewRunApproved('<Moderator_Name>', '<Player_Name>', 'Purple Duo');
 				return json({ Code: 'success' });
 			} catch (err) {
 				console.error(err.message);
@@ -719,7 +720,7 @@ export async function POST({ request }) {
 				poolConnection.close();
 
 				//returner = context.req.body;
-				notifyDiscord('<Moderator_Name>', '<Player_Name>', 'DFA Solo');
+				notifyDiscordNewRunApproved('<Moderator_Name>', '<Player_Name>', 'DFA Solo');
 				return json({ Code: 'success' });
 			} catch (err) {
 				console.error(err.message);
@@ -893,7 +894,7 @@ export async function POST({ request }) {
 				poolConnection.close();
 
 				//returner = context.req.body;
-				notifyDiscord('<Moderator_Name>', '<Player_Name>', 'Purple Solo');
+				notifyDiscordNewRunApproved('<Moderator_Name>', '<Player_Name>', 'Purple Solo');
 				return json({ Code: 'success' });
 			} catch (err) {
 				console.error(err.message);
@@ -902,42 +903,3 @@ export async function POST({ request }) {
 			break;
 	}
 }
-
-const notifyDiscord = async (moderatorName, playerName, quest) => {
-	const webhookUrl = process.env.RUN_APPROVAL_DISCORD_WEBHOOK_URL;
-	if (!webhookUrl) return;
-
-	const thumbnail = quest.toLowerCase().startsWith('dfa')
-		? process.env.RUN_APPROVAL_THUMBNAIL_DFA
-		: process.env.RUN_APPROVAL_THUMBNAIL_STIA_PURPLE;
-	try {
-		fetch(webhookUrl, {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json'
-			},
-			body: JSON.stringify({
-				username: 'PSO2 Central Leaderboard',
-				embeds: [
-					{
-						title: 'A new run has been approved :white_check_mark:',
-						url: 'https://leaderboard.pso2central.com',
-						description: `A new **${quest}** run from \`${playerName}\` has been approved by \`${moderatorName}\` and added to the leaderboard.`,
-						color: 54300,
-						thumbnail: {
-							url: process.env.RUN_APPROVAL_WEBHOOK_USER_IMG
-						},
-						image: {
-							url: thumbnail
-						},
-						footer: {
-							text: 'PSO2 Central Leaderboard'
-						}
-					}
-				]
-			})
-		});
-	} catch (err) {
-		console.error(err);
-	}
-};
